pyproject.toml
[build-system]  # https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]  # https://python-poetry.org/docs/pyproject/
name = "fitness_tracker"
version = "0.0.0"
description = "A package to group all my fitness fitness apps into one place and allow for the syncing between apps"
authors = ["Cameron Bell <cabell132@hotmail.com>"]
readme = "README.md"

[tool.commitizen]  # https://commitizen-tools.github.io/commitizen/config/
bump_message = "bump(release): v$current_version → v$new_version"
tag_format = "v$version"
update_changelog_on_bump = true
version = "0.0.0"
version_files = ["pyproject.toml:version"]


[tool.poetry.dependencies] # https://python-poetry.org/docs/dependency-specification/
pydantic = ">=1.10.7"
python = ">=3.12,<4.0"
python-dotenv = "^1.0.1"
certifi = "^2024.12.14"
sqlalchemy = "1.4.53"
alembic = "^1.14.1"
rapidfuzz = "^3.11.0"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
tqdm = "^4.67.1"
langchain = "^0.3.15"
langsmith = "^0.3.1"
langchain-openai = "^0.3.2"
dropbox = "^12.0.2"
matplotlib = "^3.10.0"
plotly = "^6.0.0"
nbformat = "^5.10.4"


[tool.poetry.group.test.dependencies] # https://python-poetry.org/docs/master/managing-dependencies/
commitizen = ">=3.2.1"
coverage = { extras = ["toml"], version = ">=7.2.5" }
mypy = ">=1.2.0"
poethepoet = ">=0.20.0"
pre-commit = ">=3.3.1"
pytest = ">=7.3.1"
pytest-clarity = ">=1.0.1"
pytest-mock = ">=3.10.0"
pytest-xdist = ">=3.2.1"
ruff = ">=0.1.3"
safety = ">=2.3.4,!=2.3.5"
shellcheck-py = ">=0.9.0"
typeguard = ">=3.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
types-sqlalchemy = "1.4.53.38"

[tool.coverage.report] # https://coverage.readthedocs.io/en/latest/config.html#report
fail_under = 50
precision = 1
show_missing = true
skip_covered = true

[tool.coverage.run]  # https://coverage.readthedocs.io/en/latest/config.html#run
branch = true
command_line = "--module pytest"
data_file = "reports/.coverage"
source = ["fitness_tracker"]

[tool.coverage.xml]  # https://coverage.readthedocs.io/en/latest/config.html#xml
output = "reports/coverage.xml"

[tool.mypy] # https://mypy.readthedocs.io/en/latest/config_file.html
junit_xml = "reports/mypy.xml"
plugins = "pydantic.mypy"
strict = true
disallow_subclassing_any = false
disallow_untyped_decorators = false
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[tool.pydantic-mypy]  # https://pydantic-docs.helpmanual.io/mypy_plugin/#configuring-the-plugin
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true


[tool.pytest.ini_options]  # https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
addopts = "--color=yes --doctest-modules --exitfirst --failed-first --strict-config --strict-markers --verbosity=2 --junitxml=reports/pytest.xml"
filterwarnings = ["error", "ignore::DeprecationWarning"]
testpaths = ["fitness_tracker", "tests"]
xfail_strict = true

[tool.ruff] # https://github.com/charliermarsh/ruff
fix = true
ignore-init-module-imports = true
line-length = 100
select = [
  "A",
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "D",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "F",
  "FBT",
  "FLY",
  "FURB",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "LOG",
  "N",
  "NPY",
  "PERF",
  "PGH",
  "PIE",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "PT",
  "PTH",
  "PYI",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TCH",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
ignore = [
  "E501",
  "RET504",
  "S101",
  "ISC001",
  "D100",
  "I001",
  "PGH003",
  "PERF203",
  "FBT001",
  "FBT002",
  "D212",
]
unfixable = ["ERA001", "F401", "F841", "T201", "T203"]
src = ["fitness_tracker", "tests"]
target-version = "py39"

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.pycodestyle]
max-doc-length = 100


[tool.ruff.pydocstyle]
convention = "google"

[tool.poe.tasks] # https://github.com/nat-n/poethepoet


[tool.poe.tasks.docs]
help = "Generate this package's docs"
cmd = """
  pdoc
    --docformat $docformat
    --output-directory $outputdirectory
    fitness_tracker
  """

[[tool.poe.tasks.docs.args]]
help = "The docstring style (default: google)"
name = "docformat"
options = ["--docformat"]
default = "google"

[[tool.poe.tasks.docs.args]]
help = "The output directory (default: docs)"
name = "outputdirectory"
options = ["--output-directory"]
default = "docs"


[tool.poe.tasks.lab]
help = "Run Jupyter Lab"

[[tool.poe.tasks.lab.sequence]]
cmd = "mkdir -p notebooks"

[[tool.poe.tasks.lab.sequence]]
cmd = """
  jupyter lab
    --allow-root
    --notebook-dir notebooks
  """

[tool.poe.tasks.lint]
help = "Lint this package"

[[tool.poe.tasks.lint.sequence]]
cmd = """
    pre-commit run
      --all-files
      --color always
    """


[[tool.poe.tasks.lint.sequence]]
shell = "safety check --continue-on-error --full-report"


[tool.poe.tasks.test]
help = "Test this package"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage run"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage report"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage xml"

[tool.semantic_release]
version_toml = ["pyproject.toml:tool.poetry.version"]
version_variables = ["fitness_tracker/__init__.py:__version__"]
version_source = "tag"
assets = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = true

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
  "build",
  "chore",
  "ci",
  "docs",
  "feat",
  "fix",
  "perf",
  "style",
  "refactor",
  "test",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false

[tool.semantic_release.remote.token]
env = "GH_TOKEN"

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
